Refactor settings page: standardize styling and reorganize components
- Apply consistent styling across the settings page for a unified look.
- Move components into appropriate packages for better project structure.
