+++ booklore-ui/src/app/metadata/book-metadata-center/book-metadata-center.component.html
@@ -1,4 +1,4 @@
-<div class="rounded-xl overflow-hidden bg-card">
+<div class="rounded-xl overflow-hidden">
   <p-tabs [value]="tab" lazy="true">
     <p-tablist>
       <p-tab value="view">Book Details</p-tab>
