+++ booklore-ui/src/app/metadata/book-metadata-center/metadata-editor/metadata-editor.component.html
@@ -22,6 +22,15 @@
         </p-progressSpinner>
 
         <div class="flex items-center gap-6">
+          <p-button
+            icon="pi pi-refresh"
+            (onClick)="regenerateCover(metadata.bookId)"
+            pTooltip="Regenerate cover"
+            tooltipPosition="top"
+            [outlined]="true"
+            [disabled]="metadata['coverLocked']">
+          </p-button>
+
           <p-fileupload
             pTooltip="Upload cover"
             tooltipPosition="top"
@@ -38,14 +47,6 @@
             [auto]="true"
             (onBeforeUpload)="onBeforeSend()">
           </p-fileupload>
-
-          <p-button
-            icon="pi pi-refresh"
-            (onClick)="regenerateCover(metadata.bookId)"
-            pTooltip="Regenerate cover"
-            tooltipPosition="top"
-            [disabled]="metadata['coverLocked']">
-          </p-button>
 
           <p-button *ngIf="!metadata['coverLocked']" icon="pi pi-lock-open" [outlined]="true" (onClick)="toggleLock('thumbnailUrl')" severity="success"></p-button>
           <p-button *ngIf="metadata['coverLocked']" icon="pi pi-lock" [outlined]="true" (onClick)="toggleLock('thumbnailUrl')" severity="warn"></p-button>
