+++ booklore-ui/src/app/metadata/book-metadata-center/metadata-searcher/metadata-searcher.component.html
@@ -16,7 +16,7 @@
     </div>
 
     <div>
-      <p-button label="Search" type="submit" [disabled]="!isSearchEnabled || loading"></p-button>
+      <p-button label="Search" type="submit" outlined="true" [disabled]="!isSearchEnabled || loading"></p-button>
     </div>
   </div>
