+++ booklore-ui/src/app/metadata/metadata-options-dialog/metadata-advanced-fetch-options/metadata-advanced-fetch-options.component.html
@@ -71,8 +71,8 @@
       </div>
     </div>
     <div class="flex flex-row gap-6">
-      <p-button severity="warn" label="Reset" (onClick)="reset()"></p-button>
-      <p-button label="Submit" (onClick)="submit()"></p-button>
+      <p-button severity="warn" outlined="true" label="Reset" (onClick)="reset()"></p-button>
+      <p-button label="Save" outlined="true" (onClick)="submit()"></p-button>
     </div>
   </div>
