+++ booklore-ui/src/app/settings/email/email-provider/email-provider.component.html
@@ -6,7 +6,7 @@
        tooltipPosition="right"
        style="cursor: pointer;"></i>
   </h2>
-  <p-button outlined="true" label="Create New Provider" (onClick)="openCreateProviderDialog()"></p-button>
+  <p-button outlined="true" icon="pi pi-plus" label="Create New Provider" (onClick)="openCreateProviderDialog()"></p-button>
 </div>
 
 <p-table [value]="emailProviders">
