+++ booklore-ui/src/app/settings/email/email-recipient/email-recipient.component.html
@@ -6,7 +6,7 @@
        tooltipPosition="right"
        style="cursor: pointer;"></i>
   </h2>
-  <p-button outlined="true" label="Add New Recipient" (onClick)="openAddRecipientDialog()"></p-button>
+  <p-button outlined="true" icon="pi pi-plus" label="Add New Recipient" (onClick)="openAddRecipientDialog()"></p-button>
 </div>
 
 <p-table [value]="recipientEmails">
