+++ booklore-ui/src/app/settings/global-preferences/file-upload-pattern/file-upload-pattern.component.scss
