+++ booklore-ui/src/app/settings/global-preferences/opds-settings/opds-settings.component.scss
