+++ booklore-ui/src/app/settings/global-preferences/opds-settings/opds-user.service.ts
@@ -1,5 +1,5 @@
 import {inject, Injectable} from '@angular/core';
-import {API_CONFIG} from './config/api-config';
+import {API_CONFIG} from '../../../config/api-config';
 import {HttpClient} from '@angular/common/http';
 import {Observable} from 'rxjs';
