+++ booklore-ui/src/app/settings/user-management/user-management.component.scss
@@ -1,0 +1,3 @@
+.enclosing-container {
+  border-color: var(--p-content-border-color);
+}
