+++ booklore-ui/src/app/settings/user-management/user-management.component.ts
@@ -5,7 +5,7 @@ import {DialogService, DynamicDialogRef} from 'primeng/dynamicdialog';
 import {CreateUserDialogComponent} from './create-user-dialog/create-user-dialog.component';
 import {TableModule} from 'primeng/table';
 import {NgIf, NgStyle} from '@angular/common';
-import {User, UserService} from '../../user.service';
+import {User, UserService} from './user.service';
 import {MessageService} from 'primeng/api';
 import {Checkbox} from 'primeng/checkbox';
 import {MultiSelect} from 'primeng/multiselect';
@@ -13,9 +13,10 @@ import {Library} from '../../book/model/library.model';
 import {LibraryService} from '../../book/service/library.service';
 import {Dialog} from 'primeng/dialog';
 import {Password} from 'primeng/password';
+import {Tooltip} from 'primeng/tooltip';
 
 @Component({
-  selector: 'app-admin',
+  selector: 'app-user-management',
   imports: [
     FormsModule,
     Button,
@@ -26,11 +27,12 @@ import {Password} from 'primeng/password';
     MultiSelect,
     Dialog,
     Password,
+    Tooltip,
   ],
-  templateUrl: './admin.component.html',
-  styleUrls: ['./admin.component.scss'],
+  templateUrl: './user-management.component.html',
+  styleUrls: ['./user-management.component.scss'],
 })
-export class AdminComponent implements OnInit {
+export class UserManagementComponent implements OnInit {
   ref: DynamicDialogRef | undefined;
   private dialogService = inject(DialogService);
   private userService = inject(UserService);
